anomaly/requirements.txt
flask
pyyaml
kafka-python